+++ plugins/auth-backend-module-oauth2-provider/config.d.ts
@@ -36,12 +36,19 @@ export interface Config {
           includeBasicAuth?: boolean;
           signIn?: {
             resolvers: Array<
-              | { resolver: 'usernameMatchingUserEntityName' }
+              | {
+                  resolver: 'usernameMatchingUserEntityName';
+                  dangerouslyAllowSignInWithoutUserInCatalog?: boolean;
+                }
               | {
                   resolver: 'emailLocalPartMatchingUserEntityName';
                   allowedDomains?: string[];
+                  dangerouslyAllowSignInWithoutUserInCatalog?: boolean;
                 }
-              | { resolver: 'emailMatchingUserEntityProfileEmail' }
+              | {
+                  resolver: 'emailMatchingUserEntityProfileEmail';
+                  dangerouslyAllowSignInWithoutUserInCatalog?: boolean;
+                }
             >;
           };
           sessionDuration?: HumanDuration | string;