+++ plugins/scaffolder/src/components/fields/EntityTagsPicker/EntityTagsPicker.test.tsx
@@ -1,0 +1,148 @@
+/*
+ * Copyright 2025 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Entity } from '@backstage/catalog-model';
+import { catalogApiRef } from '@backstage/plugin-catalog-react';
+import { catalogApiMock } from '@backstage/plugin-catalog-react/testUtils';
+import { ScaffolderRJSFFieldProps as FieldProps } from '@backstage/plugin-scaffolder-react';
+import { renderInTestApp, TestApiProvider } from '@backstage/test-utils';
+import { ComponentType, PropsWithChildren, ReactNode } from 'react';
+import { EntityTagsPicker } from './EntityTagsPicker';
+import { EntityTagsPickerProps } from './schema';
+import { useTranslationRef } from '@backstage/frontend-plugin-api';
+import { scaffolderTranslationRef } from '../../../translation';
+
+const makeEntity = (kind: string, namespace: string, name: string): Entity => ({
+  apiVersion: 'scaffolder.backstage.io/v1beta3',
+  kind,
+  metadata: { namespace, name },
+});
+
+describe('<EntityTagsPicker />', () => {
+  const entities: Entity[] = [
+    makeEntity('Group', 'default', 'team-a'),
+    makeEntity('Group', 'default', 'squad-b'),
+  ];
+  const onChange = jest.fn();
+  const schema = { type: 'array', items: { type: 'string' } };
+  const uiSchema: EntityTagsPickerProps['uiSchema'] = {
+    'ui:options': {},
+  };
+  const rawErrors: string[] = [];
+  const formData = undefined;
+
+  let props: FieldProps<string[]>;
+
+  const catalogApi = catalogApiMock.mock({
+    getEntities: jest.fn(async () => ({ items: entities })),
+  });
+
+  let Wrapper: ComponentType<PropsWithChildren<{}>>;
+
+  beforeEach(() => {
+    Wrapper = ({ children }: { children?: ReactNode }) => (
+      <TestApiProvider apis={[[catalogApiRef, catalogApi]]}>
+        {children}
+      </TestApiProvider>
+    );
+  });
+
+  afterEach(() => jest.resetAllMocks());
+
+  describe('EntityTagsPicker description', () => {
+    const description = {
+      fromSchema: 'EntityTagsPicker description from schema',
+      fromUiSchema: 'EntityTagsPicker description from uiSchema',
+    } as { fromSchema: string; fromUiSchema: string; default?: string };
+
+    beforeEach(() => {
+      const RealWrapper = Wrapper;
+      Wrapper = ({ children }: { children?: ReactNode }) => {
+        const { t } = useTranslationRef(scaffolderTranslationRef);
+        description.default = t('fields.entityTagsPicker.description');
+        return <RealWrapper>{children}</RealWrapper>;
+      };
+    });
+    it('presents default description', async () => {
+      props = {
+        onChange,
+        schema,
+        required: true,
+        uiSchema,
+        rawErrors,
+        formData,
+      } as unknown as FieldProps<any>;
+
+      const { getByText, queryByText } = await renderInTestApp(
+        <Wrapper>
+          <EntityTagsPicker {...props} />
+        </Wrapper>,
+      );
+      expect(getByText(description.default!)).toBeInTheDocument();
+      expect(queryByText(description.fromSchema)).toBe(null);
+      expect(queryByText(description.fromUiSchema)).toBe(null);
+    });
+
+    it('presents schema description', async () => {
+      props = {
+        onChange,
+        schema: {
+          ...schema,
+          description: description.fromSchema,
+        },
+        required: true,
+        uiSchema,
+        rawErrors,
+        formData,
+      } as unknown as FieldProps<any>;
+
+      const { getByText, queryByText } = await renderInTestApp(
+        <Wrapper>
+          <EntityTagsPicker {...props} />
+        </Wrapper>,
+      );
+      expect(queryByText(description.default!)).toBe(null);
+      expect(getByText(description.fromSchema)).toBeInTheDocument();
+      expect(queryByText(description.fromUiSchema)).toBe(null);
+    });
+
+    it('presents uiSchema description', async () => {
+      props = {
+        onChange,
+        schema: {
+          ...schema,
+          description: description.fromSchema,
+        },
+        required: true,
+        uiSchema: {
+          ...uiSchema,
+          'ui:description': description.fromUiSchema,
+        },
+        rawErrors,
+        formData,
+      } as unknown as FieldProps<any>;
+
+      const { getByText, queryByText } = await renderInTestApp(
+        <Wrapper>
+          <EntityTagsPicker {...props} />
+        </Wrapper>,
+      );
+      expect(queryByText(description.default!)).toBe(null);
+      expect(queryByText(description.fromSchema)).toBe(null);
+      expect(getByText(description.fromUiSchema)).toBeInTheDocument();
+    });
+  });
+});