+++ plugins/permission-node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-permission-node
 
+## 0.10.0-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-node@0.6.3-next.2
+  - @backstage/backend-plugin-api@1.3.1-next.2
+  - @backstage/config@1.3.2
+  - @backstage/errors@1.2.7
+  - @backstage/plugin-permission-common@0.9.0-next.0
+
 ## 0.10.0-next.1
 
 ### Minor Changes